requirements_colab.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=10.0.0
matplotlib>=3.6.0
seaborn>=0.12.0
tqdm>=4.64.0

# Google Colab specific
google-api-python-client>=2.0.0
google-colab>=1.0.0
kagglehub>=0.2.0

# System utilities
logging
pathlib
dataclasses
datetime
gc
json
warnings

# Optional: For enhanced features
scipy>=1.9.0
scikit-learn>=1.1.0
plotly>=5.0.0